🚚 Rename: 데이터파일 mocks 폴더로 경로 수정
