+++ src/components/auth/login/menu/list-menu.tsx
@@ -1,5 +1,5 @@
 import * as L from '../../../../styles/auth/login/login.style';
-import MenuData from '../../../../utils/auth/login/menuData';
+import MenuData from '../../../../mocks/auth/login/menuData';
 import ItemMenu from './item-menu';
 
 const ListMenu = () => {
