+++ src/components/auth/test/qna/qnaForm.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import * as T from '../../../../styles/auth/test/test.style';
 import * as S from '../../../../styles/auth/signup/signup.style';
-import QnaData from '../../../../utils/auth/test/qnaData';
+import QnaData from '../../../../mocks/auth/test/qnaData';
 import type { QnaItem } from '../../../../types/auth/test/qna';
 import { getQnaResult } from '../../../../utils/auth/test/getQnaResult';
 import ListQna from './list-qna';
