+++ src/components/footer/list-footer.tsx
@@ -1,6 +1,6 @@
 import * as F from '../../styles/footer/footer.style';
 import ItemFooter from './item-footer';
-import FooterData from '../../utils/footer/footerData';
+import FooterData from '../../mocks/footer/footerData';
 
 const ListFooter = () => {
   return (
